feat: add request models and response models, change controllers directory location
