+++ src/main/java/com/team1/efep/controllers/PageController.java
@@ -1,4 +1,4 @@
-package com.team1.efep.enums.controllers;
+package com.team1.efep.controllers;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Controller;
@@ -6,6 +6,5 @@ import org.springframework.web.bind.annotation.RequestMapping;
 
 @Controller
 @RequiredArgsConstructor
-@RequestMapping("/buyer")
-public class BuyerController {
+public class PageController {
 }
