+++ src/main/java/com/team1/efep/controllers/SellerController.java
@@ -1,4 +1,4 @@
-package com.team1.efep.enums.controllers;
+package com.team1.efep.controllers;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Controller;
@@ -6,5 +6,6 @@ import org.springframework.web.bind.annotation.RequestMapping;
 
 @Controller
 @RequiredArgsConstructor
-public class PageController {
+@RequestMapping("/seller")
+public class SellerController {
 }
