+++ src/main/java/com/team1/efep/models/request_models/AddToCartRequest.java
@@ -1,0 +1,4 @@
+package com.team1.efep.models.request_models;
+
+public class AddToCartRequest {
+}
