+++ src/main/java/com/team1/efep/models/response_models/AddToCartResponse.java
@@ -1,0 +1,4 @@
+package com.team1.efep.models.response_models;
+
+public class AddToCartResponse {
+}
